fix package dependencies and clean dead code
