+++ packages/apps/job-launcher/server/jest.config.ts
@@ -10,5 +10,6 @@ module.exports = {
   },
   moduleNameMapper: {
     '^uuid$': require.resolve('uuid'),
+    '^typeorm$': require.resolve('typeorm'),
   },
 };
