+++ packages/apps/job-launcher/server/package.json
@@ -27,19 +27,45 @@
   "dependencies": {
     "@human-protocol/core": "*",
     "@human-protocol/sdk": "*",
+    "@nestjs/axios": "^2.0.0",
     "@nestjs/common": "^9.0.0",
+    "@nestjs/config": "^2.3.4",
     "@nestjs/core": "^9.0.0",
+    "@nestjs/jwt": "^10.0.3",
+    "@nestjs/passport": "^9.0.3",
     "@nestjs/platform-express": "^9.0.0",
+    "@nestjs/schedule": "^2.2.3",
+    "@nestjs/swagger": "^6.3.0",
+    "@nestjs/terminus": "^9.2.2",
+    "@nestjs/typeorm": "^9.0.1",
+    "@types/bcrypt": "^5.0.0",
+    "@types/cookie-parser": "^1.4.3",
+    "@types/express-session": "^1.17.7",
+    "@types/passport-jwt": "^3.0.8",
+    "@types/uuid": "^9.0.2",
+    "@types/zxcvbn": "^4.4.1",
+    "bcrypt": "^5.1.0",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.0",
+    "cookie-parser": "^1.4.6",
     "ethers": "^5.7.2",
+    "helmet": "^7.0.0",
+    "nestjs-ethers": "^2.0.3",
+    "nestjs-minio-client": "^2.0.0",
+    "passport": "^0.6.0",
+    "passport-jwt": "^4.0.1",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.2.0"
+    "rxjs": "^7.2.0",
+    "typeorm": "^0.3.16",
+    "typeorm-naming-strategies": "^4.1.0",
+    "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
     "@golevelup/ts-jest": "^0.3.7",
     "@nestjs/cli": "^9.0.0",
     "@nestjs/schematics": "^9.0.0",
     "@nestjs/testing": "^9.0.0",
-    "@types/express": "^4.17.13",
+    "@types/express": "^4.17.17",
     "@types/jest": "29.5.1",
     "@types/node": "18.16.12",
     "@types/supertest": "^2.0.11",
@@ -48,6 +74,7 @@
     "eslint": "^8.0.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "express-session": "^1.17.3",
     "jest": "29.5.0",
     "prettier": "^2.3.2",
     "source-map-support": "^0.5.20",
