+++ packages/apps/job-launcher/server/src/common/constants/index.ts
@@ -1,11 +1,9 @@
 export const NS = 'hmt';
-export const COMPANY_NAME = 'Human Protocol Foundation';
 export const COINGECKO_API_URL =
   'https://api.coingecko.com/api/v3/simple/price';
 export const JOB_RETRIES_COUNT_THRESHOLD = 3;
 export const TX_CONFIRMATION_TRESHOLD = 1;
 
-export const EXCHANGE_ORACLE_ADDRESS = '';
 export const EXCHANGE_ORACLE_WEBHOOK_URL =
   'http://localhost:3005/webhook/job-launcher';
 export const RECORDING_ORACLE_ADDRESS = '';
