+++ packages/apps/job-launcher/server/src/common/enums/collection.ts
@@ -1,0 +1,4 @@
+export enum SortDirection {
+  ASC = 'ASC',
+  DESC = 'DESC',
+}
