--- packages/apps/job-launcher/server/src/common/interceptors/pagination.ts
@@ -1,15 +1,0 @@
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-import {
-  CallHandler,
-  ExecutionContext,
-  Injectable,
-  NestInterceptor,
-} from '@nestjs/common';
-
-@Injectable()
-export class PaginationInterceptor implements NestInterceptor {
-  intercept(_context: ExecutionContext, next: CallHandler): Observable<any> {
-    return next.handle().pipe(map(([rows, count]) => ({ rows, count })));
-  }
-}
