+++ packages/apps/job-launcher/server/src/common/interfaces/base.ts
