+++ packages/apps/job-launcher/server/src/common/interfaces/job.ts
@@ -1,4 +1,3 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
 import { JobStatus } from '../enums/job';
 import { IBase } from './base';
 
@@ -14,10 +13,3 @@ export interface IJob extends IBase {
   retriesCount?: number;
   waitUntil: Date;
 }
-
-export const Job = createParamDecorator(
-  (_data: unknown, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
-    return (request.job as IJob) || null;
-  },
-);
