+++ packages/apps/job-launcher/server/src/common/interfaces/user.ts
@@ -1,0 +1,9 @@
+import { UserStatus, UserType } from '../enums/user';
+import { IBase } from './base';
+
+export interface IUser extends IBase {
+  password: string;
+  email: string;
+  status: UserStatus;
+  type: UserType;
+}
