--- packages/apps/job-launcher/server/src/modules/user/user.seed.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-
-import { UserSeedService } from './user.seed.service';
-import { UserEntity } from './user.entity';
-import { UserRepository } from './user.repository';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([UserEntity])],
-  providers: [UserSeedService, UserRepository],
-  exports: [UserSeedService],
-})
-export class UserSeedModule {}
